Fix of Bug 5177 - Naive Bayes Learner & Predictor to support PMML (requires PMML 4.2).
I had to change to much code in the model classes that is why I moved the old implementation to deprecated.

The spec problem is solved however if the class column used to train the model is not a nominal value e.g. double the data field definition in the PMML data dictionary does not declare the filed as categorical but continuous. Until we can solve the problem the learner only accepts NominalValue compatible columns. However this still poses a problem with BooleanCells since they implement NominalValue as well as DoubleValue and our PMMLDataDictionaryTranslator class interprets boolean values as continuous and not categorical. I will talk to Alexander about this problem and try to set the PMML data dictionary type of the class column manually to categorical independent of its type as soon as we can do this I will change the learner to support all types PMML also supports.