AP-8977: Rewrite LogRegPredictor
The mapping from parameter to predictor is now established via an index
map calculated in the constructor of the LogRegPredictor. This change
results in a 15x faster prediction.
The predictor now also utilizes sparsity in the data by performing a
sparse matrix multiplication.

AP-8977 (Speedup Logistic Regression Predictor, especially for 'wide'
data/fingerprints (tens of hundreds of attributes))