AP-8892: RecursiveLoopEnd: Improved handling of FlowVariable edge-cases.
When configuring a Recursive Loop End node which has been just dragged
into the workbench an error was shown in the console because no
FlowVariable was available. This commit fixes this and moves the
checking for FlowVariables into onOpen(), ensuring that the available
FlowVariables are always up to date when the connections of the node
change.

AP-8892 (Recursive Loop End(s) with Option to choose FlowVariable in
Dialog)