AP-10431: Dist between miss val desc and pattern
This already solves the issue because when loading a workflow
saveSettingsTo gets called last which left the m_missValPat field with
the description of the pattern: '<empty>' or '<none>' but it should
contain the pattern '""' or 'null'. Now the m_missValPat field is not
misused anymore to store the pattern description.

Also add constants for the missing value description such that there
can't be any typos.