AP-11424: Implement different output format
Now the explanation for each row comprises of multiple rows where each
row contains the explanation for a single target.
This commit also introduces the concept of Explanation as a dedicated
class that should be reused by LIME and SHAP in the future.

AP-11424 (Implement Shapley Values for Model Interpretability)