AP-13569: Changed encoding selection method for the line reader node
Moved the encoding selection to an extra tab in the line reader dialog.
The already for the csv reader used class "CharsetNamePanel" is used
now, but the design was slightly adjusted according to this tickets
specification. After the revision, the name of the default charset is
now displayed in brackets.

AP-13569 (Revise charset selection dialog component (layout and errors))