AP-13433: Implemented shared classes for filehandling util
Implemented a swing worker which check if the include parent folder
can be checked based on a entered path in the filechooser. This
will be used at least in the copy move file and uncompress node. Also
added a Path manipulator interface as well as a new class which
implements this interface to relativzize a file path based on several
options.

AP-13433 (Copy / move files no table input (new file handling))