AP-13433: Fixed copy/move desc and folder bug
Fixed the node description to match the functionality and fixed the bug
that the node always creates folders which are non existent even if the
create missing folders option is unchecked. Also removed the
functionalty that if the missing folder option is unchecked and a single
non existent folder is added to an existing path the
StatusMessageReporter gives an error as well if the node gets executed.

AP-13433 (Copy / move files no table input (new file handling))