AP-15382: Copy complete folder Copy Move
Added the option to completely copy a folder with all its files and
folders (including empty folders). As well as a new option to let the
node fail if something goes wrong during the deletion process.

AP-15382 (Copy/Move files ensure that empty folder are being copied when
Folder is selected)