AP-15641: CompressionAwareCountingInputStream
Introduced a class which wraps an input stream in an countinginputstream
or when necessary in a GZIPInputStream and holds the CountingStream and
the InputStream. This class will be used for instance in the csvreader
and linereader which are able to read .gz files to get the correct
progress.

AP-15641 (Get correct file size of .gz files)