AP-15768: Add enforce types option to transformation table
If selected, it is attempted to find an alternative production path to
the configured KNIME type. Should no such path exist, the node will fail
with a corresponding error message.
If not selected, the default production path is used for the new
external type.

AP-15768 (Type mapping re-constructing production paths changes node
behavior)