AP-15781: Don't hide SettingsModelFileSystem in connected case
This was a bug because it introduced a name conflict with settings
stored by the node framework and also caused problems when
adding/removing the file system port.
The StringToPathNodeModel has been adapted to load settings stored in
4.2 correctly. This only affects the case where an fs port is present
because nothing changes otherwise. The fix is to not validate/load the
settings in this case which isn't a problem because they are overwritten
by the port object anyway.

AP-15781 (String to Path (Variable) (New File Handling))