AP-16353: Fixed csv writer column header bug
Fixed the problem that when the overwrite policy is not append and the
"dont write column header" option is enabled but greyed out the option
is still active. The option should only be used if the overwrite policy
is append.

AP-16353 (CSV Writer does not write column header when overwriting and
disabled "don't write column header" option is enabled)