AP-15819: Improve validation of URI format settings
There is an issue with some DialogComponents (such as DialogComponentDuration)
which only transfer the value from the UI to their SettingsModel during
saveSettings(), but not before. Hence we need to first save() settings
and then validate them (instead of the other way around).

AP-15819 (Add support for signed URLs to Amazon S3 Connector)