AP-16704: Allow buffering of preview for certain readers
For some readers, reading a single row is very expensive (e.g. JSON and Table Reader).
The current preview implementation fetches the rows in the EDT causing the UI to freeze.
As a workaround we introduce the possibility to first buffer the preview table into a
BufferedDataTable before the preview is displayed. The loading is done in the background
and a progress bar is displayed.

AP-16704 (Expensive RowIterators cause table view to freeze)