AP-17365: Fixed TRF renaming bug
Fixed a bug in the TRF which caused a problem when a user tries to
rename a column to a name from an unselected column and vice versa.
Fixed another problem which is caused by not resetting the
transformation names when reseting the names in the dialog.

AP-17365 (Filtered columns are still considered as duplicate column
names)