AP-13269: Fix error when not using output cols
This fixes an exception when no output columns are
selected. The error occurs only before the dialog
has columns included for the first time, then it does
not occur anymore. This is due to the fact that initially
the array is null, after adding and removing a column
it is not null anymore.

AP-13269 (Value Lookup (supersedes "Cell Replacer" node with functionality similar to Excel's VLOOKUP))