AP-20669: Fix implementation of Column Rename
Previously the node casted columns to string and only ensured
compatibility afterwards.
This is not possible with the ColumnarBackend because it has to find the
correct ValueFactory for each column.
Therefore, the conversion to string now happens before the type cast and
it is also made sure that the domain of the converted string columns
(i.e. their possible values) are preserved.
Note that conversion only happens if the target type is string, so if an
attempt is made to cast a column to any other incompatible type, the
cast will still fail. However, in such cases the row-based backend would
also fail in subsequent nodes.