UIEXT-1836: Make field in ColumnFilter optional
The custom persistor only works on settings that contain the respective key. This is correct and would also lead to no problems in the ColumnFilter if the NodeModel was a WebUiNodeModel (since then the settings when the dialog is opened are saved from DefaultNodeSettings already.
UIEXT-1836 (InvalidSettingsException when opening dialog for unconnected Column Filter)