AP-22914: Row Filter to show ordering-based operators only for "bounded&numeric" data types
This splits former "enabled for" check into "applicable" and "offered", where the former is used during execute
to check whether we want to apply the operator to the input data, and the latter is used to determine whether to offer
the operator in the dialog. It can still be supplied via a flow variable and existing instances still execute,
though they will show a "warning" in the dialog, e.g. the "(MISSING) LT" option.

AP-22914 (Row Filter unexpectedly offers "ordering" comparison operators for some data types)