AP-22914: Keep previously input value
Since we don't "reset" the operator anymore when the column is missing or incompatible, we need to keep the current
value with the previously selected columns' type. When the user switches back to a column of the same type, the value
can be used again.

AP-22914 (Row Filter unexpectedly offers "ordering" comparison operators for some data types)