AP-23194: AP-23194: Re-apply last fix for 5.3.3 inclusion ("findPattern" parameter in String Replacer)
This reverts commit eacacc5ea184480ca98fab1098580be1e5edf4b8.

This commit reverts the last undo commit as the fix was not meant to go
into 5.3.2. That release is now out the door, so this fix can go onto
the branch for inclusion in 5.3.3.

AP-23194 (Importing WF from AP <= 5.0 to AP 5.1, 5.2 or 5.3 with a
string replacer causes parameters to not appear)
AP-23194 (Importing WF from AP <= 5.0 to AP 5.1, 5.2 or 5.3 with a string replacer causes parameters to not appear)