AP-23606: Quick fix K-AI row filtering with a single bound
Optional arguments should be handled by the K-AI node func framework,
but currently we receive the string "null" if an argument is left empty,
instead of an Object that is null. This is a quick fix to handle such
cases for the row filter.

AP-23606 (K-AI Build Mode: Row Filter NodeFuncs don't work properly)