AP-23017: Fix hard cast on `BooleanCell` and allow more column types
The boolean predicate factory is now consistent with the other predicate factories,
in that it checks for the preferred value class instead of `instanceof BooleanCell`.

The fix for the columnar backend is to change the cast from `BooleanCell` to `BooleanValue`,
but this change also enables use of contributed boolean types.

AP-23017 (Row Filter to allow equality check between compatible numeric types)