UIEXT-2609: Revert "UIEXT-2609: Load default for 'updateDomain' field"
This reverts commit 246f20eab63f4f58911d87e75c1e2c2e317570b2.

We did not have to backport the changes from 5.3 to 5.2 after all since the "washing" of model settings ensured that the settings defaults from the model were loaded before opening the dialog.
We moved the "washing to when legacy NodeDialogPanes" are being opened with 5.3 to enable deprecation of individual config keys without breaking flow variables.

UIEXT-2609 (Reference Row filter configuration resets to default from legacy to webUI dialog)