UIEXT-2923: Make it possible to apply stash from legacy settings
This has been possible until we removed the serializers from DynamicValuesInput (which was necessary to not have to use jackson in knime-base and also they included a lot of code that is not needed anymore)
But the state provider mechanism collects dependencies from json (even for computeBeforeOpenDialog), so the settings we use still need to be serializable.
So I simply use a class with one String field now and defer the loading to DynamicValuesInput to whenever we make use of them.

UIEXT-2923 (Use dynamic parameters in the RowFilter)