UIEXT-1740: Introduce flexible modernUI reader framework
I.e. a framework that allows keeping the existing node model implementations depending on AbstractMultiTableReadConfig but filling that config from NodeParameters.
This make the first version of webui readers obsolote which is now deprecated.
It's not easily possible to migrate an existing reader to this framework without deprecation. Since this also impacts execution behavior, we should not do that.

Next to the common reader classes and the implementations for CSV and Knime Table reader this commit also provides a tutorial with steps to follow for other readers.
For now this tutorial is focused on creating a webUI dialog for a reader which already has a config.

In one of the next improvements, the following things will be added:
* Support also for MultiFileSelection
* The ReaderParameters will be split into very small pieces so that we do not enforce on any reader that these settings have to be part of it. It should be modular and e.g. one should be able to easily replace Single- by Multiple-FileSelection

The current code has the following known and accepted (since not yet needed) limitation(s):
* The default parameters are constructed with empty specs (i.e. no reader should depend on the input specs in its constructor).

UIEXT-1740 (Migrate Node Model of reader nodes)