UIEXT-3029: Use correct table spec for pmml output
Orienting on how it is done in the DB-Version of this node but using a better variable name as the newDataTableSpec there. It is not the output spec that we need to use but the input spec plus the appended columns (otherwise applying the global transformation fails).

UIEXT-3029 (Binner does not work with DB Apply-Binner)